+++ src/django_aiogram/eventlog/writer.py
@@ -23,7 +23,7 @@ from django.utils import timezone
 
 from django_aiogram.eventlog.dbrouter import event_log_database
 from django_aiogram.eventlog.events import short_id
-from django_aiogram.eventlog.recorder import Event
+from django_aiogram.eventlog.records import Event
 from django_aiogram.exceptions import DjangoRedisAiogramError
 from django_aiogram.models import TelegramEvent
 from django_aiogram.wire.payloads import redact_keys, redact_text, redact_values, secrets